Fix consecutive device select issue
Selecting the same device consecutively may cause a problem; it is not only redundant.
This is because device selection may send <UCP>[Power], and dual sending of the command finally turn the device off.
Change to invalidate the second device selection for the same device if the action for the first selection is still ongoing.

Bug: 379739489
Change-Id: I7ce99c615495546e6f0b90bbe2737b819121211c