Revert "Change filter logic to allow Context Hub owners"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce85d06a-c919-4cfb-b5ea-1e079feb61ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e1d7b02c6ccfc18639a861ed769f8f2b0df3aea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com