Detect weak charger and shutdown at 0% battery
Summary: Before this change, the device would shutdown at 0% battery if no charger was plugged in. This logic worked most of the time for shutting down the device at 0% battery.

There is a corner case here though. When the battery drops close to 0%, some users are still actively using the device. They ignore the low battery warning, plug in a charger and keep using the device. Plugging in any charger will skip the shutdown logic. But because the charger isn't supplying enough wattage to actually charge the battery, the battery voltage will get drained to well below 3V and trigger the UVLO protection. Hitting UVLO will shutdown the device abruptly with no visual/audio notification to the users, which is not a good user experience.

This change alters the behavior above. If the battery is at 0% and a weak charger is detected, then proceed with a software initiated shutdown.

Test: Successful Build on master branch

Change-Id: I4dd6eca88e516f58b1916ba7f06478b13766451f
Signed-off-by: Abdelrahman Daim <adaim@meta.com>