Introduce a SystemFeaturesMetadata class
This is a simple generated helper class with the # of defined system
features, as computed by an annotation processor for
@SdkConstant(SdkConstantType.FEATURE).

For now, we use this as a size hint to the system feature IPC cache. A
follow-up change will augment the codegen such that we can compute a
"perfect" indexing of SDK system features to indices, which we can use
to optimize system feature queries without paying a high per-process
memory cost for a runtime, string-based hash.

Bug: 203143243
Test: m
Change-Id: I979a14711ec8554a7b7b4e900ab78b38321b8b93