Revert "aconfig: Add min_sdk for NFC module dependencies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d3adf3d2-6e07-4486-8944-5d4884cbedb6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382493210
Change-Id: I14f472384bbe8c46eda0c6ad188db37cd4e592d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com