Solve the issue where the pip menu flickers after the screen lights up and the first click on pip.
This is a workaround; the scenario for reproducing the problem is
that after clicking on pip, the pip menu is displayed. At this point,
if the screen is immediately turned off and then turned back on, and
pip is clicked, there will be a flicker. The original issue was that
after the screen was turned off, the surface where the pip menu is
located was set to invisible by its parent layer, but when the screen
was turned on, the last frame before the screen was turned off would
immediately display, and then the alpha animation would be executed.
Here, we directly force the corresponding surface to be visible or
invisible during the show and hide operations of the pip menu.

Test: manual -Pip
Change-Id: Ie9f1231054adb28591f682cc695c0e881115f936